@@ -11,57 +11,61 @@
     "ionic:serve": "ionic-app-scripts serve"
 ***REMOVED***,
   "dependencies": {
-    "@angular/common": "4.4.4",
-    "@angular/compiler": "4.4.4",
-    "@angular/compiler-cli": "4.4.4",
-    "@angular/core": "4.4.4",
-    "@angular/forms": "4.4.4",
-    "@angular/http": "4.4.4",
-    "@angular/platform-browser": "4.4.4",
-    "@angular/platform-browser-dynamic": "4.4.4",
-    "@firebase/webchannel-wrapper": "^0.2.1",
-    "@ionic-native/core": "4.3.2",
-    "@ionic-native/file": "4.3.2",
-    "@ionic-native/file-opener": "4.3.2",
-    "@ionic-native/network": "4.3.2",
-    "@ionic-native/splash-screen": "4.3.2",
-    "@ionic-native/status-bar": "4.3.2",
-    "@ionic/storage": "^2.0.1",
-    "@types/c3": "^0.4.42",
-    "@types/leaflet": "1.0.63",
-    "@types/papaparse": "4.1.28",
-    "angularfire2": "5.0.0-rc.3",
-    "c3": "^0.4.12",
-    "cordova-android": "^6.2.3",
-    "cordova-android-support-gradle-release": "^1.1.2",
+    "@angular/common": "6.0.4",
+    "@angular/compiler": "6.0.4",
+    "@angular/compiler-cli": "6.0.4",
+    "@angular/core": "6.0.4",
+    "@angular/forms": "6.0.4",
+    "@angular/http": "6.0.4",
+    "@angular/platform-browser": "6.0.4",
+    "@angular/platform-browser-dynamic": "6.0.4",
+    "@firebase/webchannel-wrapper": "^0.2.8",
+    "@ionic-native/core": "4.7.0",
+    "@ionic-native/file": "4.7.0",
+    "@ionic-native/file-opener": "4.7.0",
+    "@ionic-native/network": "4.7.0",
+    "@ionic-native/splash-screen": "4.7.0",
+    "@ionic-native/status-bar": "4.7.0",
+    "@ionic/storage": "^2.1.3",
+    "angularfire2": "5.0.0-rc.10",
+    "c3": "^0.6.2",
+    "cordova-android": "^7.1.0",
+    "cordova-android-support-gradle-release": "^1.4.3",
     "cordova-plugin-compat": "^1.2.0",
     "cordova-plugin-console": "^1.1.0",
-    "cordova-plugin-device": "^1.1.6",
-    "cordova-plugin-file": "^4.3.3",
+    "cordova-plugin-device": "^2.0.2",
+    "cordova-plugin-file": "^6.0.1",
     "cordova-plugin-file-opener2": "^2.0.19",
-    "cordova-plugin-network-information": "^1.3.3",
-    "cordova-plugin-splashscreen": "^4.0.3",
-    "cordova-plugin-statusbar": "^2.2.3",
-    "cordova-plugin-whitelist": "^1.3.2",
-    "cordova-sqlite-storage": "^2.0.4",
-    "firebase": "^4.5.0",
-    "ionic-angular": "3.8.0",
+    "cordova-plugin-network-information": "^2.0.1",
+    "cordova-plugin-splashscreen": "^5.0.2",
+    "cordova-plugin-statusbar": "^2.4.2",
+    "cordova-plugin-whitelist": "^1.3.3",
+    "cordova-sqlite-storage": "^2.3.2",
+    "firebase": "^5.0.4",
+    "ionic-angular": "3.9.2",
     "ionic-plugin-keyboard": "^2.2.1",
-    "ionicons": "3.0.0",
-    "leaflet": "1.0.3",
+    "ionicons": "4.1.2",
+    "leaflet": "1.3.1",
     "leaflet-ajax": "^2.1.0",
-    "ng2-canvas-whiteboard": "^1.2.0",
-    "papaparse": "^4.3.3",
-    "pdfjs-dist": "^1.8.446",
-    "promise-polyfill": "6.0.2",
-    "rxjs": "5.4.3",
+    "ng2-canvas-whiteboard": "^1.4.5",
+    "papaparse": "^4.5.0",
+    "pdfjs-dist": "^2.0.489",
+    "promise-polyfill": "8.0.0",
+    "rxjs": "6.2.0",
+    "rxjs-compat": "^6.2.0",
     "sw-toolbox": "3.6.0",
-    "zone.js": "0.8.18"
+    "zone.js": "0.8.26"
 ***REMOVED***,
   "devDependencies": {
-    "@ionic/app-scripts": "^3.0.0-201710161925",
-    "@types/request": "0.0.30",
-    "typescript": "2.3.4"
+    "@ionic/app-scripts": "^3.1.10",
+    "@types/c3": "^0.6.0",
+    "@types/leaflet": "1.2.7",
+    "@types/papaparse": "4.1.34",
+    "@types/request": "2.47.0",
+    "prettier": "^1.13.5",
+    "tslint": "^5.10.0",
+    "tslint-config-prettier": "^1.13.0",
+    "typescript": "2.9.1"
 ***REMOVED***,
   "cordovaPlugins": [
     "cordova-plugin-whitelist",
