@@ -17,29 +17,29 @@ firebase could use email format 'localID@picsa'
 note - haven't included all merge features, should review when testing
 
 */
-import { Injectable } from '@angular/core';
-import { Http } from '@angular/http';
-import { ToastController, Events } from 'ionic-angular'
-import { Platform } from "ionic-angular";
-import { Storage } from '@ionic/storage';
-import { FileOpener } from '@ionic-native/file-opener';
-import { File } from '@ionic-native/file';
-import { AngularFirestore } from 'angularfire2/firestore';
-
-// auth
+import { Injectable } from "@angular/core";
+import { Http } from "@angular/http";
+import { File } from "@ionic-native/file";
+import { FileOpener } from "@ionic-native/file-opener";
+import { Storage } from "@ionic/storage";
 import { AngularFireAuth } from "angularfire2/auth";
-import * as firebase from 'firebase/app';
+import { AngularFirestore } from "angularfire2/firestore";
+import * as firebase from "firebase/app";
+import { Events, ToastController } from "ionic-angular";
+import { Platform } from "ionic-angular";
+// auth
 
-import 'rxjs/add/operator/map';
+import { map } from "rxjs/operators";
+import { constants } from "fs";
 
 @Injectable()
 export class StorageProvider {
-  public user: any
-  userID: string
-  firebaseRef
-  initCalled: boolean = false
-  online: boolean = false
-  firebaseID:string
+  public user: any;
+  userID: string;
+  firebaseRef;
+  initCalled: boolean = false;
+  online: boolean = false;
+  firebaseID: string;
   // lots of constructor code can be cleaned up after migration to newer version
   constructor(
     public http: Http,
@@ -49,84 +49,88 @@ export class StorageProvider {
     public platform: Platform,
     public events: Events,
     private file: File,
-    private afs: AngularFirestore,
-  ) {
-***REMOVED***
+    private afs: AngularFirestore
+  ) {}
 
   storageInit() {
     this.storage.ready().then(_ => {
-      this.initCalled = true
-      this._checkDB().then((upgraded) => {
+      this.initCalled = true;
+      this._checkDB().then(upgraded => {
         if (upgraded) {
           // database up to date and user exists
-          this.loadUser().then((user) => {
-            this.events.publish('user:loaded', user)
-        ***REMOVED***)
-      ***REMOVED***
-        else {
+          this.loadUser().then(user => {
+            this.events.publish("user:loaded", user);
+        ***REMOVED***);
+      ***REMOVED*** else {
           this._migrateData().then(() => {
             this.getUser().then(() => {
-              return
-          ***REMOVED***)
-        ***REMOVED***)
+              return;
+          ***REMOVED***);
+        ***REMOVED***);
       ***REMOVED***
-    ***REMOVED***)
-  ***REMOVED***)
-
+    ***REMOVED***);
+  ***REMOVED***);
 ***REMOVED***
 
   loadUser() {
     return new Promise((resolve, reject) => {
-      this.storage.get('userID').then(id => {
+      this.storage.get("userID").then(id => {
         if (id) {
-          this.userID = id
-          this.getUserDoc('settings', 'profile').then(user => {
-            console.log('user doc retrieved', user)
-            this.user = user
-            resolve(this.user)
-        ***REMOVED***)
-      ***REMOVED***
-        else {
-          console.log('no user,creating')
+          this.userID = id;
+          this.getUserDoc("settings", "profile").then(user => {
+            console.log("user doc retrieved", user);
+            this.user = user;
+            resolve(this.user);
+        ***REMOVED***);
+      ***REMOVED*** else {
+          console.log("no user,creating");
           this._createUser().then(user => {
-            this.user = user
-            resolve(this.user)
-        ***REMOVED***)
+            this.user = user;
+            resolve(this.user);
+        ***REMOVED***);
       ***REMOVED***
-    ***REMOVED***)
-  ***REMOVED***)
+    ***REMOVED***);
+  ***REMOVED***);
 ***REMOVED***
 
   getUser() {
     // checks for user existance within local this.storage.
     //returns corresponding firestore user doc if exists, or creates new if not
     return new Promise((resolve, reject) => {
-      if (this.user) { resolve(this.user) }
-      else {
-        if (!this.initCalled) { this.storageInit() }
+      if (this.user) {
+        resolve(this.user);
+    ***REMOVED*** else {
+        if (!this.initCalled) {
+          this.storageInit();
+      ***REMOVED***
         // try again in a moment. alternatively could use event emitters but not sure if there would still potentially be gap between setting up and init finishing
-        this.events.subscribe('user:loaded', (user) => {
-          this.events.unsubscribe('user:loaded')
-          resolve(user)
-      ***REMOVED***)
+        this.events.subscribe("user:loaded", user => {
+          this.events.unsubscribe("user:loaded");
+          resolve(user);
+      ***REMOVED***);
         setTimeout(() => {
-          this.getUser().then(() => resolve(this.user))
-      ***REMOVED***, 5000
-        )
+          this.getUser().then(() => resolve(this.user));
+      ***REMOVED***, 5000);
     ***REMOVED***
-  ***REMOVED***)
+  ***REMOVED***);
 ***REMOVED***
 
   getUserDoc(collection, docId?) {
     // ***need another function to return from local db
     // ***could also add queries
-    let userID = this.userID
+    const userID = this.userID;
     return new Promise((resolve, reject) => {
       // local first approach
-      this.storage.get(collection).then(res => {
-        if(docId){resolve(res[docId])}
-        else{resolve(res)}
-    ***REMOVED***).catch(err=>console.log('failed retrieving user doc',err))
+      this.storage
+        .get(collection)
+        .then(res => {
+          if (docId) {
+            resolve(res[docId]);
+        ***REMOVED*** else {
+            resolve(res);
+        ***REMOVED***
+      ***REMOVED***)
+        .catch(err => console.log("failed retrieving user doc", err));
       // this.afs.firestore.collection("users").doc(this.userID).collection(collection).doc(docId).get()
       //   .then(res => {
       //     //*** should use res to check for exist before running data */
@@ -141,56 +145,64 @@ export class StorageProvider {
       //     resolve({})
       // ***REMOVED***)
       //   .catch(err => console.log('could not get doc', collection, docId, userID, err))
-  ***REMOVED***)
+  ***REMOVED***);
 ***REMOVED***
 
   removeUserDoc(collection, docId) {
-    console.log('removing user doc', collection, docId)
+    console.log("removing user doc", collection, docId);
     // offline first
     return new Promise((resolve, reject) => {
       this.storage.get(collection).then(res => {
         if (res.hasOwnProperty(docId)) {
-          delete res[docId]
+          delete res[docId];
       ***REMOVED***
-        this.storage.set(collection, res)
-        resolve()
-    ***REMOVED***)
-  ***REMOVED***)
+        this.storage.set(collection, res);
+        resolve();
+    ***REMOVED***);
+  ***REMOVED***);
     // return this.afs.firestore.collection("users").doc(this.userID).collection(collection).doc(docId).delete()
 ***REMOVED***
 
   getAll(collection) {
     // get all docs in collection, returns as array
-    console.log('getting all docs in collection', collection)
+    console.log("getting all docs in collection", collection);
     return new Promise((resolve, reject) => {
       // local first approach
-      this.storage.get(collection).then((res) => {
-        console.log('docs received, converting to array', res)
+      this.storage.get(collection).then(res => {
+        console.log("docs received, converting to array", res);
         // messy with lots of try/catches because of old and mixed data storage
         // will be removed in future
-        try { res=JSON.parse(res)}
-        catch(err) { }
-        if (res == null) { resolve([]) }
-        else {
-          let docsArray = []
+        try {
+          res = JSON.parse(res);
+      ***REMOVED*** catch (err) {}
+        if (res == null) {
+          resolve([]);
+      ***REMOVED*** else {
+          const docsArray = [];
           for (let key in res) {
             if (res.hasOwnProperty(key)) {
-              let data = res[key]
-              if (data.hasOwnProperty('jsonString')) {
-                try {data = JSON.parse(data.jsonString)}
-                catch(err) {data = data.jsonString}
+              let data = res[key];
+              if (data.hasOwnProperty("jsonString")) {
+                try {
+                  data = JSON.parse(data.jsonString);
+              ***REMOVED*** catch (err) {
+                  data = data.jsonString;
+              ***REMOVED***
             ***REMOVED***
-              if (data.hasOwnProperty('json')) {
-                try {data = JSON.parse(data.json)}
-                catch(err) {data = data.json}
+              if (data.hasOwnProperty("json")) {
+                try {
+                  data = JSON.parse(data.json);
+              ***REMOVED*** catch (err) {
+                  data = data.json;
+              ***REMOVED***
             ***REMOVED***
-              docsArray.push(data)
-              console.log('array', docsArray)
-              resolve(docsArray)
+              docsArray.push(data);
+              console.log("array", docsArray);
+              resolve(docsArray);
           ***REMOVED***
         ***REMOVED***
       ***REMOVED***
-    ***REMOVED***)
+    ***REMOVED***);
       // this.afs.firestore.collection("users").doc(this.userID).collection(collection).get().then(snapshot => {
       //   console.log('snapshot', snapshot)
       //   // convert json back to string and push into array
@@ -205,86 +217,110 @@ export class StorageProvider {
       //   console.log('docs array', docsArray)
       //   resolve(docsArray)
       // })
-  ***REMOVED***)
-
+  ***REMOVED***);
 ***REMOVED***
 
   getSharedDoc() {
     // get doc not saved in own user profile
-
     // should change getDoc to getUserDoc
 ***REMOVED***
 
-
-  saveUserDoc(data: any, stringify: boolean, collection?: string, id?: string, merge?: boolean) {
+  saveUserDoc(
+    data: any,
+    stringify: boolean,
+    collection?: string,
+    id?: string,
+    merge?: boolean
+  ) {
     // saves data attached to user profile
     // accepts data, whether to stringify (avoid nested arrays), optional colletion and document id
-    console.log('saving', data, stringify, collection, id)
+    console.log("saving", data, stringify, collection, id);
 
     return new Promise((resolve, reject) => {
       if (!this.userID) {
-        return this.loadUser().then(() => this.saveUserDoc(data, stringify, collection, id, merge))
+        return this.loadUser().then(() =>
+          this.saveUserDoc(data, stringify, collection, id, merge)
+        );
     ***REMOVED***
-      if (stringify == true) { data = { jsonString: JSON.stringify(data) } }
-      if (!merge) { merge = false }
+      if (stringify == true) {
+        data = { jsonString: JSON.stringify(data) ***REMOVED***
+    ***REMOVED***
+      if (!merge) {
+        merge = false;
+    ***REMOVED***
 
       if (collection) {
         // create new doc within collection, overrides any previous document
         // offline first approach
         // *** note ,used to be based on returning functions but now promise so live needs to swap for resolve methods ***
-        if (!id) { id = this.generatePushID() }
-        console.log('creating new doc in collection by id', collection, id)
+        if (!id) {
+          id = this.generatePushID();
+      ***REMOVED***
+        console.log("creating new doc in collection by id", collection, id);
         this.storage.get(collection).then(
           res => {
-            if (res == null) { res = {} }
-            res[id] = data
-            this.storage.set(collection, res).then(_ => resolve(res))
+            if (res == null) {
+              res = {***REMOVED***
+          ***REMOVED***
+            res[id] = data;
+            this.storage.set(collection, res).then(_ => resolve(res));
         ***REMOVED***,
           rej => {
-            this.storage.set(collection, { id: data }).then(_ => resolve({ id: data }))
-        ***REMOVED***)
+            this.storage
+              .set(collection, { id: data })
+              .then(_ => resolve({ id: data }));
+        ***REMOVED***
+        );
 
         // if (id) { return this.afs.firestore.collection('users').doc(this.userID).collection(collection).doc(id).set(data, { merge: merge }) }
         // else { return this.afs.firestore.collection('users').doc(this.userID).collection(collection).doc().set(data, { merge: merge }) }
-    ***REMOVED***
-      else {
+    ***REMOVED*** else {
         // otherwise update any existing fields, uses set command with merge option to prevent total overwrite
         //offline first approach
         this.storage.get(collection).then(res => {
           if (merge) {
-            for (let key in data) {
-              res[key] = data[key]
+            for (const key in data) {
+              res[key] = data[key];
           ***REMOVED***
+        ***REMOVED*** else {
+            res = data;
         ***REMOVED***
-          else { res = data }
-          this.storage.set(collection, res).then(_ => { resolve(res) })
-      ***REMOVED***)
+          this.storage.set(collection, res).then(_ => {
+            resolve(res);
+        ***REMOVED***);
+      ***REMOVED***);
         // return this.afs.firestore.collection('users').doc(this.userID).set(data, { merge: merge })
     ***REMOVED***
-
-  ***REMOVED***)
-
+  ***REMOVED***);
 ***REMOVED***
 
-  saveBatch(data: any, stringify: boolean, collection: string, idAsKey?: boolean) {
-    console.log('saving batch', data)
-    console.log('this.userID', this.userID)
+  saveBatch(
+    data: any,
+    stringify: boolean,
+    collection: string,
+    idAsKey?: boolean
+  ) {
+    console.log("saving batch", data);
+    console.log("this.userID", this.userID);
     // save multiple docs to a sub collection on user doc. overwrites any existing doc
     // allows idAsKey if data format {id:data}, which will maintain same doc ref
 
     // offline first
-    let temp = {}
-    for (let key in data) {
-      let doc = data[key]
+    const temp = {***REMOVED***
+    for (const key in data) {
+      let doc = data[key];
       if (stringify) {
-        doc = { json: JSON.stringify(doc) }
+        doc = { json: JSON.stringify(doc) ***REMOVED***
     ***REMOVED***
-      let id
-      if (idAsKey) { id = key }
-      else { id = this.generatePushID() }
-      temp[id] = doc
+      let id;
+      if (idAsKey) {
+        id = key;
+    ***REMOVED*** else {
+        id = this.generatePushID();
+    ***REMOVED***
+      temp[id] = doc;
   ***REMOVED***
-    return this.storage.set(collection, temp)
+    return this.storage.set(collection, temp);
     // let batch = this.afs.firestore.batch();
     // for (let key in data) {
     //   // create key:value pair doc
@@ -304,129 +340,135 @@ export class StorageProvider {
     // return batch.commit()
 ***REMOVED***
 
-  updateForm(submissionID){
-    // 
+  updateForm(submissionID) {
+    //
 ***REMOVED***
 
-  syncForms(firebaseID){
-    return new Promise((resolve,reject)=>{
-      this.storage.get('submittedForms').then(forms=>{
+  syncForms(firebaseID) {
+    return new Promise((resolve, reject) => {
+      this.storage.get("submittedForms").then(forms => {
         // hardcoded just for reporting forms for now
-        let pending = forms.reporting.pending
-        let batch = this.afs.firestore.batch();
-        console.log('pending',pending)
-        for(let p of pending){
-          let id=p._submissionID
-          let ref = this.afs.firestore.collection('forms').doc('reporting').collection('submissions').doc(id)
-          batch.set(ref,p)
+        const pending = forms.reporting.pending;
+        const batch = this.afs.firestore.batch();
+        console.log("pending", pending);
+        for (let p of pending) {
+          let id = p._submissionID;
+          let ref = this.afs.firestore
+            .collection("forms")
+            .doc("reporting")
+            .collection("submissions")
+            .doc(id);
+          batch.set(ref, p);
       ***REMOVED***
-        batch.commit().then(
-          res => resolve('success'),
-          rej => reject(rej)
-        ).catch(err=>console.log('problem syncing forms',err))
-    ***REMOVED***)
-  ***REMOVED***)
-   
+        batch
+          .commit()
+          .then(res => resolve("success"), rej => reject(rej))
+          .catch(err => console.log("problem syncing forms", err));
+    ***REMOVED***);
+  ***REMOVED***);
 ***REMOVED***
 
   syncAll(firebaseID) {
     // assumes preflight checks already carried out in network app
     return new Promise((resolve, reject) => {
-      console.log('syncing all', firebaseID)
-      let batch = this.afs.firestore.batch();
-      let total = 0
-      this.storage.forEach((docsObject, collection, n) => {
-        /* storage in form
+      console.log("syncing all", firebaseID);
+      const batch = this.afs.firestore.batch();
+      let total = 0;
+      this.storage
+        .forEach((docsObject, collection, n) => {
+          /* storage in form
         collection       docs object
         budgets:{ budgets:data1, budget2:data2    ***REMOVED***
         */
 
-        // push collections objects to right place
-        if (typeof docsObject == 'object') {
-          for (let id in docsObject) {
-            let data = docsObject[id]
-            if(data instanceof Date){
-              // custom format for date object instance
-              data ={***REMOVED*** data[id]=docsObject
+          // push collections objects to right place
+          if (typeof docsObject == "object") {
+            for (let id in docsObject) {
+              let data = docsObject[id];
+              if (data instanceof Date) {
+                // custom format for date object instance
+                data = {***REMOVED***
+                data[id] = docsObject;
+            ***REMOVED***
+              const ref = this.afs.firestore
+                .collection("users")
+                .doc(firebaseID)
+                .collection(collection)
+                .doc(id);
+              batch.set(ref, data);
+              total++;
           ***REMOVED***
-            let ref = this.afs.firestore.collection('users').doc(firebaseID).collection(collection).doc(id)
-            batch.set(ref, data)
-            total++
         ***REMOVED***
-      ***REMOVED***
-        //anything else is local setting, can still pass
-        else {
-          let ref = this.afs.firestore.collection('users').doc(firebaseID)
-          let data = {}
-          data[collection] = docsObject
-          batch.set(ref, data)
-          total++
-      ***REMOVED***
-    ***REMOVED***).then(_ => {
-        console.log('commiting', total)
-        console.log('batch',batch)        
-        batch.commit()
-          .then(
-            res => resolve('success'),
-            rej => reject(rej))
-          .catch(err => console.log('err', err));
-    ***REMOVED***)
-  ***REMOVED***)
-    
-
+          //anything else is local setting, can still pass
+          else {
+            const ref = this.afs.firestore.collection("users").doc(firebaseID);
+            const data = {***REMOVED***
+            data[collection] = docsObject;
+            batch.set(ref, data);
+            total++;
+        ***REMOVED***
+      ***REMOVED***)
+        .then(_ => {
+          console.log("commiting", total);
+          console.log("batch", batch);
+          batch
+            .commit()
+            .then(res => resolve("success"), rej => reject(rej))
+            .catch(err => console.log("err", err));
+      ***REMOVED***);
+  ***REMOVED***);
 ***REMOVED***
 
   _createUser(id?) {
     //creates new user and saves to this.user
     return new Promise((resolve, reject) => {
-      if (!id) { id = this.generatePushID() }
+      if (!id) {
+        id = this.generatePushID();
+    ***REMOVED***
       this.user = {
         id: id,
-        name: 'anonymous',
-        role: 'extension',
-        group: 'malawi-2017'
-    ***REMOVED***
-      this.userID = id
-      this.events.publish('user:loaded', this.user)
+        name: "anonymous",
+        role: "extension",
+        group: "malawi-2017"
+    ***REMOVED***;
+      this.userID = id;
+      this.events.publish("user:loaded", this.user);
       // save id to local storage and sync to firebase db (offline and online)
-      this.storage.set('userID', this.userID).then(() => {
-        this.saveUserDoc(this.user, false, 'settings', 'profile')
-          .then(() => {
-            console.log('user saved', this.user);
-            resolve(this.user)
-        ***REMOVED***)
-    ***REMOVED***)
-  ***REMOVED***)
+      this.storage.set("userID", this.userID).then(() => {
+        this.saveUserDoc(this.user, false, "settings", "profile").then(() => {
+          console.log("user saved", this.user);
+          resolve(this.user);
+      ***REMOVED***);
+    ***REMOVED***);
+  ***REMOVED***);
 ***REMOVED***
 
   _checkDB() {
-    return this.storage.get('dbUpgraded')
+    return this.storage.get("dbUpgraded");
 ***REMOVED***
 
- 
-
-
   assignPermissions(code) {
-    console.log('assigning permissions')
+    console.log("assigning permissions");
     return new Promise((resolve, reject) => {
-      this.loadFile('assets/admin/userPermissions.json').then(res => {
+      this.loadFile("assets/admin/userPermissions.json").then(res => {
         if (res[code]) {
-          this.user.permissions = res[code]
-          this.saveUserDoc(this.user, false, 'settings', 'profile')
-            .then(_ => resolve(this.user))
+          this.user.permissions = res[code];
+          this.saveUserDoc(this.user, false, "settings", "profile").then(_ =>
+            resolve(this.user)
+          );
+      ***REMOVED*** else {
+          reject("Invalid code, please try again");
       ***REMOVED***
-        else {
-          reject('Invalid code, please try again')
-      ***REMOVED***
-    ***REMOVED***)
-  ***REMOVED***)
+    ***REMOVED***);
+  ***REMOVED***);
 ***REMOVED***
 
   loadFile(url) {
-    var options = {}
+    var options = {***REMOVED***
     return new Promise(resolve => {
-      this.http.get(url)
-        .map(res => res.json())
+      this.http
+        .get(url)
+        .pipe(map(res => res.json()))
         .subscribe(data => {
           resolve(data);
       ***REMOVED***);
@@ -434,29 +476,31 @@ export class StorageProvider {
 ***REMOVED***
 
   presentToast(message) {
-    console.log('creating toast', message)
-    let toast = this.toastCtrl.create({
+    console.log("creating toast", message);
+    const toast = this.toastCtrl.create({
       message: message,
       duration: 3000
   ***REMOVED***);
     toast.present();
 ***REMOVED***
 
-
   generatePushID() {
-    var PUSH_CHARS = '-0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZ_abcdefghijklmnopqrstuvwxyz';
-    var lastPushTime = 0;
-    var lastRandChars = [];
-    var now = new Date().getTime();
-    var duplicateTime = (now === lastPushTime);
+    const PUSH_CHARS =
+      "-0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZ_abcdefghijklmnopqrstuvwxyz";
+    let lastPushTime = 0;
+    const lastRandChars = [];
+    let now = new Date().getTime();
+    const duplicateTime = now === lastPushTime;
     lastPushTime = now;
-    var timeStampChars = new Array(8);
+    const timeStampChars = new Array(8);
     for (var i = 7; i >= 0; i--) {
       timeStampChars[i] = PUSH_CHARS.charAt(now % 64);
       now = Math.floor(now / 64);
   ***REMOVED***
-    if (now !== 0) throw new Error('We should have converted the entire timestamp.');
-    var id = timeStampChars.join('');
+    if (now !== 0) {
+      throw new Error("We should have converted the entire timestamp.");
+  ***REMOVED***
+    let id = timeStampChars.join("");
     if (!duplicateTime) {
       for (i = 0; i < 12; i++) {
         lastRandChars[i] = Math.floor(Math.random() * 64);
@@ -470,7 +514,9 @@ export class StorageProvider {
     for (i = 0; i < 12; i++) {
       id += PUSH_CHARS.charAt(lastRandChars[i]);
   ***REMOVED***
-    if (id.length != 20) throw new Error('Length should be 20.');
+    if (id.length != 20) {
+      throw new Error("Length should be 20.");
+  ***REMOVED***
     return id;
 ***REMOVED***
 
@@ -479,111 +525,111 @@ export class StorageProvider {
     // will be removed for future deployments
     return new Promise((resolve, reject) => {
       this.storage.keys().then(keys => {
-        console.log('keys', keys)
+        console.log("keys", keys);
         if (keys.length == 0) {
           // new user
-          console.log('new user')
-          this._createUser().then((user) => {
-            this.storage.set('dbUpgraded', true)
-            resolve(user)
-        ***REMOVED***)
-
-      ***REMOVED***
-        else {
+          console.log("new user");
+          this._createUser().then(user => {
+            this.storage.set("dbUpgraded", true);
+            resolve(user);
+        ***REMOVED***);
+      ***REMOVED*** else {
           // first upgrade user, then budgets
-          if (keys.indexOf('user') > -1) {
-            this._upgradeUser().then((userID) => {
-              console.log('user id', userID)
-              if (keys.indexOf('budgets') > -1) {
-                console.log('upgrading budgets')
+          if (keys.indexOf("user") > -1) {
+            this._upgradeUser().then(userID => {
+              console.log("user id", userID);
+              if (keys.indexOf("budgets") > -1) {
+                console.log("upgrading budgets");
                 this._upgradeBudgets(userID).then(() => {
-                  if (keys.indexOf('budgetCards') > -1) {
-                    console.log('upgrading budget cards')
+                  if (keys.indexOf("budgetCards") > -1) {
+                    console.log("upgrading budget cards");
                     this._upgradeBudgetCards(userID).then(() => {
-                      resolve()
-                  ***REMOVED***)
+                      resolve();
+                  ***REMOVED***);
+                ***REMOVED*** else {
+                    resolve();
                 ***REMOVED***
-                  else { resolve() }
-              ***REMOVED***)
+              ***REMOVED***);
+            ***REMOVED*** else {
+                resolve();
             ***REMOVED***
-              else { resolve() }
-          ***REMOVED***)
+          ***REMOVED***);
+        ***REMOVED*** else {
+            resolve();
         ***REMOVED***
-          else resolve()
       ***REMOVED***
-    ***REMOVED***)
-  ***REMOVED***)
-    // 
-
+    ***REMOVED***);
+  ***REMOVED***);
+    //
 
     // list db objects, then map
 ***REMOVED***
   _upgradeBudgets(userID) {
-    console.log('user id', userID)
+    console.log("user id", userID);
     return new Promise((resolve, reject) => {
-      this.storage.get('budgets').then(res => {
-        let budgets = JSON.parse(res)
-        console.log('parsed budgets', budgets)
-        this.saveBatch(budgets, true, 'budgets', true).then(() => {
-          console.log('budgets saved');
-          resolve()
-      ***REMOVED***)
-    ***REMOVED***)
-  ***REMOVED***)
+      this.storage.get("budgets").then(res => {
+        const budgets = JSON.parse(res);
+        console.log("parsed budgets", budgets);
+        this.saveBatch(budgets, true, "budgets", true).then(() => {
+          console.log("budgets saved");
+          resolve();
+      ***REMOVED***);
+    ***REMOVED***);
+  ***REMOVED***);
 ***REMOVED***
   _upgradeBudgetCards(userID) {
     // identical to above, could have been merged
-    console.log('user id', userID)
+    console.log("user id", userID);
     return new Promise((resolve, reject) => {
-      this.storage.get('budgetCards').then(res => {
-        let cards = JSON.parse(res)
-        console.log('parsed cards', cards)
-        this.saveBatch(cards, true, 'budgetCards', true).then(() => {
-          console.log('budgetCards saved');
-          resolve()
-      ***REMOVED***)
-    ***REMOVED***)
-  ***REMOVED***)
+      this.storage.get("budgetCards").then(res => {
+        const cards = JSON.parse(res);
+        console.log("parsed cards", cards);
+        this.saveBatch(cards, true, "budgetCards", true).then(() => {
+          console.log("budgetCards saved");
+          resolve();
+      ***REMOVED***);
+    ***REMOVED***);
+  ***REMOVED***);
 ***REMOVED***
   _upgradeUser() {
-    console.log('upgrading user')
+    console.log("upgrading user");
     return new Promise((resolve, reject) => {
       // existing user in need of migrating
-      let id: string
-      this.storage.get('user').then(user => {
-        console.log('user', user)
+      let id: string;
+      this.storage.get("user").then(user => {
+        console.log("user", user);
         try {
           user = JSON.parse(user);
       ***REMOVED*** catch (err) {
           // user parsed, return format either object or string depending on version. Now to upgrade
-          if (typeof user == 'string') {
+          if (typeof user == "string") {
             // 0.27 format where user saved as single string object
-            console.log('string format')
-            id = user
-            console.log('user id', user)
+            console.log("string format");
+            id = user;
+            console.log("user id", user);
         ***REMOVED***
           //else console.error(err);
       ***REMOVED***
-        console.log('parsed user', user)
-        if (user.hasOwnProperty('ID')) {
+        console.log("parsed user", user);
+        if (user.hasOwnProperty("ID")) {
           // 0.28 format with user ID field
-          id = user.ID
+          id = user.ID;
       ***REMOVED***
-        if (user.hasOwnProperty('id')) {
+        if (user.hasOwnProperty("id")) {
           // 0.29 correct format, only needs syncing as done for both below
-          id = user.id
+          id = user.id;
       ***REMOVED***
-        console.log('id', id)
-        this.storage.set('userID', this.userID)
-        this.storage.set('dbUpgraded', true)
+        console.log("id", id);
+        this.storage.set("userID", this.userID);
+        this.storage.set("dbUpgraded", true);
         this._createUser(id).then(user => {
-          console.log('this.userID', this.userID)
-          this.storage.set('dbUpgraded', true).then(() => {
-            resolve(id)
-        ***REMOVED***)
-      ***REMOVED***)
-    ***REMOVED***)
-  ***REMOVED***)
+          console.log("this.userID", this.userID);
+          this.storage.set("dbUpgraded", true).then(() => {
+            resolve(id);
+        ***REMOVED***);
+      ***REMOVED***);
+    ***REMOVED***);
+  ***REMOVED***);
 ***REMOVED***
 
   // sync(data, collection?) {
@@ -625,7 +671,6 @@ export class StorageProvider {
   //     // let db = this.afoDatabase.list('/users/' + this.userID)
   //     // console.log('db',db)
 
-
   //     // promise.offline.then(() => console.log('offline data saved to device storage!'));
 
   //     // promise.then(() => {
@@ -643,23 +688,37 @@ export class StorageProvider {
   //can merge code from resources page to single provider (either storage or file)
   //checks for a single directory (assumes picsa directory will already exist)...not adapted for root eg. /picsa/backups/profile/...
   checkFileDirectory(dir?) {
-    console.log('checking dir', dir)
-    console.log('cordova?', this.platform.is('cordova'))
-    if (!this.platform.is('cordova')) { return }
+    console.log("checking dir", dir);
+    console.log("cordova?", this.platform.is("cordova"));
+    if (!this.platform.is("cordova")) {
+      return;
+  ***REMOVED***
     return new Promise((resolve, reject) => {
-      //assumes directory child of picsa, check picsa exists 
-      this.file.checkDir(this.file.externalApplicationStorageDirectory + 'picsa/', dir)
+      //assumes directory child of picsa, check picsa exists
+      this.file
+        .checkDir(this.file.externalApplicationStorageDirectory + "picsa/", dir)
         .then(_ => {
-          console.log('directory exists', this.file.externalApplicationStorageDirectory + 'picsa/' + dir)
-          resolve('directory exists')
+          console.log(
+            "directory exists",
+            this.file.externalApplicationStorageDirectory + "picsa/" + dir
+          );
+          resolve("directory exists");
       ***REMOVED***)
         .catch(err => {
-          this.file.createDir(this.file.externalApplicationStorageDirectory + 'picsa/', dir, false).then(() => {
-            console.log('picsa/' + dir + ' directory created')
-            resolve('directory created')
-        ***REMOVED***).catch(err => { reject(err) })
-      ***REMOVED***)
-  ***REMOVED***)
+          this.file
+            .createDir(
+              this.file.externalApplicationStorageDirectory + "picsa/",
+              dir,
+              false
+            )
+            .then(() => {
+              console.log("picsa/" + dir + " directory created");
+              resolve("directory created");
+          ***REMOVED***)
+            .catch(err => {
+              reject(err);
+          ***REMOVED***);
+      ***REMOVED***);
+  ***REMOVED***);
 ***REMOVED***
-
 }
