@@ -1,34 +1,33 @@
-import { NgModule, ErrorHandler } from '@angular/core';
-import { IonicApp, IonicModule, IonicErrorHandler } from 'ionic-angular';
-import { BrowserModule } from '@angular/platform-browser';
-import { HttpModule, Http } from '@angular/http';
-import { IonicStorageModule } from '@ionic/storage';
-import { File } from '@ionic-native/file';
+/* tslint:disable:ordered-imports */
+import { NgModule, ErrorHandler } from "@angular/core";
+import { IonicApp, IonicModule, IonicErrorHandler } from "ionic-angular";
+import { BrowserModule } from "@angular/platform-browser";
+import { HttpModule, Http } from "@angular/http";
+import { IonicStorageModule } from "@ionic/storage";
+import { File } from "@ionic-native/file";
 // Ionic native modules
-import { SplashScreen } from '@ionic-native/splash-screen';
+import { SplashScreen } from "@ionic-native/splash-screen";
 
-import { Network} from '@ionic-native/network';
-import { PdfViewerComponent } from 'ng2-pdf-viewer';
-import { CanvasWhiteboardModule } from 'ng2-canvas-whiteboard';
-import { FileOpener } from '@ionic-native/file-opener';
+import { Network } from "@ionic-native/network";
+import { PdfViewerComponent } from "ng2-pdf-viewer";
+import { CanvasWhiteboardModule } from "ng2-canvas-whiteboard";
+import { FileOpener } from "@ionic-native/file-opener";
 // App pages
-import { MyApp } from './app.component';
+import { MyApp } from "./app.component";
 // Angular firestore
-import { AngularFireModule } from 'angularfire2';
-import { AngularFirestoreModule } from 'angularfire2/firestore';
-import { AngularFireAuthModule } from 'angularfire2/auth';
-import { environment } from '../environments/environment';
+import { AngularFireModule } from "angularfire2";
+import { AngularFirestoreModule } from "angularfire2/firestore";
+import { AngularFireAuthModule } from "angularfire2/auth";
+import { environment } from "../environments/environment";
 // Providers
-import {KoboApi} from "../providers/kobo-api";
 // import {FileService} from "../providers/file-service";
-import {YoutubeService} from '../providers/youtube-service';
-import { C3ChartProvider } from '../providers/c3-chart/c3-chart';
-import { ForumServiceProvider } from '../providers/forum-service/forum-service';
-import { MalawiDataProvider } from '../providers/c3-chart/malawi-data';
-import { BudgetToolProvider } from '../providers/budget-tool/budget-tool';
-import { StorageProvider } from '../providers/storage/storage';
-import { NetworkProvider } from '../providers/network/network';
-
+import { YoutubeService } from "../providers/youtube-service";
+import { C3ChartProvider } from "../providers/c3-chart/c3-chart";
+import { ForumServiceProvider } from "../providers/forum-service/forum-service";
+import { MalawiDataProvider } from "../providers/c3-chart/malawi-data";
+import { BudgetToolProvider } from "../providers/budget-tool/budget-tool";
+import { StorageProvider } from "../providers/storage/storage";
+import { NetworkProvider } from "../providers/network/network";
 
 // AF2 Settings
 export const firebaseConfig = {
@@ -40,9 +39,7 @@ export const firebaseConfig = {
 ***REMOVED***
 
 @NgModule({
-  declarations: [
-    MyApp,
-  ],
+  declarations: [MyApp],
   imports: [
     IonicModule.forRoot(MyApp),
     AngularFireModule.initializeApp(firebaseConfig),
@@ -50,19 +47,17 @@ export const firebaseConfig = {
     AngularFirestoreModule.enablePersistence(),
     AngularFireAuthModule,
     IonicStorageModule.forRoot({
-      name: '__picsa',
+      name: "__picsa"
   ***REMOVED***),
     BrowserModule,
     HttpModule,
     CanvasWhiteboardModule
   ],
   bootstrap: [IonicApp],
-  entryComponents: [
-    MyApp,
-  ],
+  entryComponents: [MyApp],
 
-  providers: [{ provide: ErrorHandler, useClass: IonicErrorHandler },
-    KoboApi,
+  providers: [
+    { provide: ErrorHandler, useClass: IonicErrorHandler },
     YoutubeService,
     // FileService,
     SplashScreen,
@@ -75,7 +70,6 @@ export const firebaseConfig = {
     FileOpener,
     File,
     NetworkProvider
-    ]
-
+  ]
 })
 export class AppModule {}
