@@ -1,14 +1,20 @@
-import { Component } from '@angular/core';
-import { IonicPage, NavController, NavParams, ViewController } from 'ionic-angular';
-import * as L from 'leaflet';
-import 'leaflet-ajax';
-import { MalawiDataProvider } from '../../../../providers/c3-chart/malawi-data';
-
+import { Component } from "@angular/core";
+import {
+  IonicPage,
+  NavController,
+  NavParams,
+  ViewController
+} from "ionic-angular";
+import * as L from "leaflet";
+import "leaflet-ajax";
+import malawiGeo1 from "../../../../assets/geoJson/Malawi-admin-1";
+import malawiGeo2 from "../../../../assets/geoJson/Malawi-admin-2-Blantyre";
+import { MalawiDataProvider } from "../../../../providers/c3-chart/malawi-data";
 
 @IonicPage()
 @Component({
-  selector: 'site-select',
-  templateUrl: 'site-select.html',
+  selector: "site-select",
+  templateUrl: "site-select.html"
 })
 export class SiteSelectPage {
   map: any;
@@ -17,76 +23,73 @@ export class SiteSelectPage {
     public navCtrl: NavController,
     public navParams: NavParams,
     public viewCtrl: ViewController,
-    private malawiData: MalawiDataProvider)
-  { }
+    private malawiData: MalawiDataProvider
+  ) {}
 
   ionViewDidLoad() {
-
-    this.map = L.map('siteSelect', {
+    this.map = L.map("siteSelect", {
       attributionControl: false
   ***REMOVED***);
-    var osmUrl = 'http://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png';
-    var osmAttrib = 'Map data © <a href="http://openstreetmap.org">OpenStreetMap</a> contributors';
-    var osm = new L.TileLayer(osmUrl, { minZoom: 1, maxZoom: 15, attribution: osmAttrib });
+    const osmUrl = "http://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png";
+    const osmAttrib =
+      'Map data © <a href="http://openstreetmap.org">OpenStreetMap</a> contributors';
+    const osm = new L.TileLayer(osmUrl, {
+      minZoom: 1,
+      maxZoom: 15,
+      attribution: osmAttrib
+  ***REMOVED***);
     // start the map in South-East England
-    this.map.setView(new L.LatLng(-13.70, 33.21), 6);
+    this.map.setView(new L.LatLng(-13.7, 33.21), 6);
     // this.map.addLayer(osm);
     //cast L to any as typings sometimes not working
 
-    // let geojsonLayer:any = new (<any>L).geoJson.ajax("assets/geoJson/Malawi-admin-2-Blantyre.geojson", {
-    let geojsonLayer: any = new (<any>L).geoJson.ajax("assets/geoJson/Malawi-admin-1.geojson", {
+    const geojsonLayer = L.geoJSON(malawiGeo1, {
       onEachFeature: this.setFeature.bind(this),
-      middleware: function (data) {
-        return data
+      middleware: function(data) {
+        return data;
     ***REMOVED***,
       // filter:filterFunction
       style: this._getStyle()
-  ***REMOVED***)
-    geojsonLayer.addTo(this.map)
+  ***REMOVED***);
+    geojsonLayer.addTo(this.map);
     // console.log('geojson bounds',geojsonLayer.getBounds())
-    this.map.fitBounds([
-      [-16.01463, 34.71830], [-15.34875, 35.13236]
-    ])
+    this.map.fitBounds([[-16.01463, 34.7183], [-15.34875, 35.13236]]);
     // this.map.on('click',function(e){
     //   console.log('clicked')
     // })
     this.malawiData.getMetaData().then(res => {
+      const sites = res["sites"];
+      for (const site of sites) {
+        console.log("site", site);
+        const marker = L.marker([site.latitude, site.longitude]);
 
-      var sites = res['sites']
-      for (let site of sites) {
-        console.log('site', site)
-        var marker = L.marker([site.latitude, site.longitude])
-
-        var container = L.DomUtil.create('div');
-        var btn = L.DomUtil.create('button', '', container)
-        btn.setAttribute('type', 'button');
-        btn.innerHTML = '<div>'+site.name+'<br>Load Data</div>';
-        var popup = L.popup().setContent(btn)
-        L.DomEvent.on(btn, 'click', (btn) => {
-          this.selectSite(site)
+        const container = L.DomUtil.create("div");
+        const btn = L.DomUtil.create("button", "", container);
+        btn.setAttribute("type", "button");
+        btn.innerHTML = "<div>" + site.name + "<br>Load Data</div>";
+        const popup = L.popup().setContent(btn);
+        L.DomEvent.on(btn, "click", btn => {
+          this.selectSite(site);
       ***REMOVED***);
-        marker.bindPopup(popup)
-        marker.addTo(this.map)
+        marker.bindPopup(popup);
+        marker.addTo(this.map);
         marker.on({
-          click: function (e) {
-            console.log('marker clicked', e)
+          click: function(e) {
+            console.log("marker clicked", e);
         ***REMOVED***.bind(this)
-      ***REMOVED***)
+      ***REMOVED***);
     ***REMOVED***
   ***REMOVED***);
-
-
-
 ***REMOVED***
   markerClick() {
-    console.log('marker clicked')
+    console.log("marker clicked");
 ***REMOVED***
 
   setFeature(feature, layer) {
-    var exceptions = {
+    const exceptions = {
       "TA Kapeni": [-15.60583, 35.00381],
       "TA Machinjili": [-15.67858, 35.07111]
-  ***REMOVED***
+  ***REMOVED***;
     layer.on({
       // mouseover: function () {
       //     this.setStyle({
@@ -98,9 +101,9 @@ export class SiteSelectPage {
       //         'fillColor': '#f0d1b1',
       //   ***REMOVED***);
       // },
-      click: function (e) {
+      click: function(e) {
         // console.log('feature',feature)
-        console.log('e', e)
+        console.log("e", e);
         // e.target.setStyle({
         //     'fillColor': '#80FF9F',
         //     'fillOpacity':1
@@ -110,22 +113,28 @@ export class SiteSelectPage {
         //possible add code to recolour previous points
         //e.g. https://stackoverflow.com/questions/25773389/changing-the-style-of-each-feature-in-a-leaflet-geojson-layer
     ***REMOVED***.bind(this)
-  ***REMOVED***)
-
+  ***REMOVED***);
 
     //automatically bind tooltips to centre of feature, unless want to manually specify from exceptions
     if (!exceptions[feature.properties.NAME_1]) {
-      layer.bindTooltip(feature.properties.NAME_1, { permanent: true, direction: 'center', className: 'countryLabel' });
-  ***REMOVED***
-    else {
-      var latLon = exceptions[feature.properties.NAME_1]
-      var label = L.marker(latLon, {
+      layer.bindTooltip(feature.properties.NAME_1, {
+        permanent: true,
+        direction: "center",
+        className: "countryLabel"
+    ***REMOVED***);
+  ***REMOVED*** else {
+      const latLon = exceptions[feature.properties.NAME_1];
+      const label = L.marker(latLon, {
         icon: L.divIcon({
-          html: '',
+          html: "",
           iconSize: [0, 0]
       ***REMOVED***)
-    ***REMOVED***).addTo(this.map)
-      label.bindTooltip(feature.properties.NAME_1, { permanent: true, direction: 'center', className: 'countryLabel' });
+    ***REMOVED***).addTo(this.map);
+      label.bindTooltip(feature.properties.NAME_1, {
+        permanent: true,
+        direction: "center",
+        className: "countryLabel"
+    ***REMOVED***);
   ***REMOVED***
 ***REMOVED***
 
@@ -133,16 +142,13 @@ export class SiteSelectPage {
     return {
       fillColor: "#f0d1b1",
       fillOpacity: 1,
-      color: '#000000',
+      color: "#000000",
       opacity: 1,
       weight: 2
-  ***REMOVED***
+  ***REMOVED***;
 ***REMOVED***
   selectSite(site) {
     // let data = e.target.feature.properties
     this.viewCtrl.dismiss(site);
 ***REMOVED***
-
-
-
 }
